JBIDE-25386: Consolidate tests from 'org.hibernate.eclipse.console.test' into new test plugin 'org.jboss.tools.hibernate.orm.test' - Move class 'org.hibernate.eclipse.console.test.project.xpl.JavaProjectHelper' to plugin 'org.hibernate.eclipse.jdt.ui.test'
